Refactor and add varying category_ids
